// graphql/generated/mutations/calendarEvent.mutations.ts
/* Do not edit this file. It was generated programmatically. */

import { CalendarEventCreationArgs } from "@/graphql/generated/args/calendarEvent.args";
import {
  calendarEventFragment,
  CalendarEventFragment,
} from "@/graphql/generated/fragments/calendarEvent.fragment";
import {
  CalendarEventCreationInput,
  CalendarEventUpdateInput,
} from "@/graphql/generated/inputs/calendarEvent.inputs";
import { gql, MutationHookOptions } from "@apollo/client";

export const CREATE_CALENDAR_EVENT = gql`
  mutation CreateCalendarEvent($data: CalendarEventCreationInput!) {
    createCalendarEvent(data: $data) {
      ...CalendarEventFragment
    }
  }
  ${calendarEventFragment}
`;

export const getOptimisticResponseForCalendarEventCreation = (
  data: CalendarEventCreationInput
): { createCalendarEvent: CalendarEventFragment } => {
  const now = new Date();
  return {
    createCalendarEvent: {
      __typename: "CalendarEvent",
      id: "tmp-id",
      remoteId: null,
      scheduleId: null,
      habitId: null,
      taskId: null,
      end: null,
      allDay: false,
      notes: null,
      canceled: null,
      archivedAt: null,
      ...data,
      createdAt: now,
      updatedAt: now,
    },
  };
};

export const updateCacheAfterCreatingCalendarEvent: MutationHookOptions<
  { createCalendarEvent: CalendarEventFragment },
  CalendarEventCreationArgs
> = {
  update(cache, { data }) {
    const { createCalendarEvent } = data || {};
    if (createCalendarEvent) {
      const newCalendarEventRef = cache.writeFragment({
        data: createCalendarEvent,
        fragment: gql`
          fragment NewCalendarEvent on CalendarEvent {
            ...CalendarEventFragment
          }
          ${calendarEventFragment}
        `,
        fragmentName: "NewCalendarEvent",
      });
      cache.modify({
        id: `User:${createCalendarEvent.userId}`,
        fields: {
          calendarEvents(existingCalendarEventRefs = []) {
            return [...existingCalendarEventRefs, newCalendarEventRef];
          },
        },
      });
      cache.modify({
        fields: {
          calendarEvents(existingCalendarEvents = []) {
            return [...existingCalendarEvents, newCalendarEventRef];
          },
        },
      });
    }
  },
};

export const UPDATE_CALENDAR_EVENT = gql`
  mutation UpdateCalendarEvent(
    $where: CalendarEventWhereUniqueInput!
    $data: CalendarEventUpdateInput!
  ) {
    updateCalendarEvent(where: $where, data: $data) {
      ...CalendarEventFragment
    }
  }
  ${calendarEventFragment}
`;

export const getOptimisticResponseForCalendarEventUpdate = (
  fragment: CalendarEventFragment,
  data: CalendarEventUpdateInput
) => {
  const now = new Date();
  return {
    updateCalendarEvent: {
      __typename: "CalendarEvent",
      ...fragment,
      ...data,
      updatedAt: now,
    },
  };
};

export const UPSERT_CALENDAR_EVENT = gql`
  mutation UpdateCalendarEvent(
    $where: CalendarEventWhereUniqueInput!
    $data: CalendarEventCreationInput!
  ) {
    upsertCalendarEvent(where: $where, data: $data) {
      ...CalendarEventFragment
    }
  }
  ${calendarEventFragment}
`;

export const UPDATE_CALENDAR_EVENTS = gql`
  mutation UpdateCalendarEvents(
    $where: CalendarEventWhereInput!
    $data: CalendarEventUpdateInput!
  ) {
    updateCalendarEvents(where: $where, data: $data) {
      ...CalendarEventFragment
    }
  }
  ${calendarEventFragment}
`;

export const UPDATE_CALENDAR_EVENTS_DISTINCTLY = gql`
  mutation UpdateCalendarEventsDistinctly($data: [CalendarEventUpdateInput!]!) {
    updateCalendarEventsDistinctly(data: $data) {
      ...CalendarEventFragment
    }
  }
  ${calendarEventFragment}
`;